Complete restructure of app structure
We're now using Rack::URLMap to define the paths for each "controller" and have
each in their own separate class. Calling OpenS3.app(path, token) should return
the rack application that can be spawned wherever.

In addition to all this, the default token is no longer set and needs to be
handled separately as no defaults are in place.